OnRuntimeUpgrade struct & Shibuya cleanup
